Changes done. 7 june

+++ lib/Screens/Distributor/DataModel.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 class DataModel{
   String title;
   String image;
-  double counts;
+  int counts;
   Color color;
 
   DataModel(this.title, this.image, this.counts,this.color);
